refactor: Use script TypeScript settings
