+++ .eslintignore
@@ -14,3 +14,5 @@ coverage
 **/*.generated.ts
 /tools/dist
 /patches
+
+.eslintrc.js
