+++ package.json
@@ -29,9 +29,10 @@
     "tslib": "2.3.1"
   },
   "devDependencies": {
+    "@renovate/eslint-plugin": "https://github.com/renovatebot/eslint-plugin#v0.0.4",
     "@types/jest": "26.0.24",
-    "@typescript-eslint/eslint-plugin": "4.33.0",
-    "@typescript-eslint/parser": "4.33.0",
+    "@typescript-eslint/eslint-plugin": "5.9.0",
+    "@typescript-eslint/parser": "5.9.0",
     "babel-jest": "26.6.3",
     "babel-plugin-dynamic-import-node": "2.3.3",
     "codecov": "3.8.3",
@@ -39,6 +40,7 @@
     "eslint-config-airbnb-typescript": "12.3.1",
     "eslint-config-prettier": "8.3.0",
     "eslint-plugin-import": "2.25.4",
+    "eslint-plugin-only-warn": "1.0.3",
     "eslint-plugin-promise": "4.3.1",
     "jest": "26.6.3",
     "jest-junit": "12.3.0",
