+++ tsconfig.json
@@ -8,12 +8,17 @@
     "outDir": "./dist",
     "target": "es2018",
     "moduleResolution": "node",
-    "importHelpers": true
+    "importHelpers": true,
+    "strict": true,
+    "noImplicitReturns": true,
+    "noFallthroughCasesInSwitch": true,
+    "noUncheckedIndexedAccess": true,
+    "noImplicitOverride": true,
+    "noUnusedLocals": true
   },
   "typedocOptions": {
     "mode": "modules",
     "out": "docs"
   },
-  "include": ["lib/**/*"],
-  "exclude": ["node_modules", "dist"]
+  "include": ["lib/**/*.ts"]
 }
